+++ .gitignore
@@ -12,4 +12,4 @@ dist/
 yalc.lock
 .rpt2_cache/
 package/
-integration_tests/benchmarks/bundle.js
+integration_tests/benchmarks/ui/bundle.js
