+++ integration_tests/benchmarks/package.json
@@ -10,7 +10,6 @@
   "devDependencies": {
     "@tensorflow/tfjs": "~0.12.4",
     "@types/jasmine": "~2.5.53",
-    "browserstack-runner": "~0.7.0",
     "clang-format": "~1.2.2",
     "jasmine": "~3.1.0",
     "jasmine-core": "~3.1.0",
@@ -29,12 +28,9 @@
     "http-server": "^0.11.1"
   },
   "scripts": {
-    "build": "tsc && browserify *.ts -p [ tsify ] > bundle.js",
     "lint": "tslint -p . -t verbose",
-    "server": "yarn build && http-server",
-    "test": "karma start --firebaseKey $FIREBASE_KEY",
-    "test-browserstack": "browserstack-runner",
-    "test-travis": "./test_travis.sh"
+    "benchmark": "karma start --firebaseKey $FIREBASE_KEY",
+    "benchmark-travis": "./test_travis.sh"
   },
   "license": "Apache-2.0",
   "engines": {
