+++ integration_tests/benchmarks/ui/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig",
   "include": [
-    "*.ts"
-  ]
+    "**/*.ts"
+  ],
+  "exclude": []
 }
